base de datos

+++ app/src/main/java/com/example/actividad/studentDataBase/model/Student.kt
@@ -17,5 +17,5 @@ data class Student(
     @ColumnInfo(name = "grupo")
     val grupo: Char,
     @ColumnInfo(name = "promedio")
-    val puntaje: Long
+    val puntaje: Float
 )
